Switched to 32-bit variables when possible, removed unneeded casts Thanks @Fix94 for the tip

+++ source/crypto.c
@@ -57,7 +57,7 @@ __asm__\
 void aes_setkey(u8 keyslot, const void* key, u32 keyType, u32 mode)
 {
 	if(keyslot <= 0x03) return; // Ignore TWL keys for now
-	u32* key32 = (u32*)key;
+	u32 * key32 = (u32 *)key;
 	*REG_AESCNT = (*REG_AESCNT & ~(AES_CNT_INPUT_ENDIAN | AES_CNT_INPUT_ORDER)) | mode;
 	*REG_AESKEYCNT = (*REG_AESKEYCNT >> 6 << 6) | keyslot | AES_KEYCNT_WRITE;
 
@@ -78,7 +78,7 @@ void aes_use_keyslot(u8 keyslot)
 
 void aes_setiv(const void* iv, u32 mode)
 {
-	const u32* iv32 = (const u32*)iv;
+	const u32 *iv32 = (const u32 *)iv;
 	*REG_AESCNT = (*REG_AESCNT & ~(AES_CNT_INPUT_ENDIAN | AES_CNT_INPUT_ORDER)) | mode;
 
 	// Word order for IV can't be changed in REG_AESCNT and always default to reversed
@@ -98,9 +98,9 @@ void aes_setiv(const void* iv, u32 mode)
 	}
 }
 
-void aes_advctr(void* ctr, u32 val, u32 mode)
+void aes_advctr(void *ctr, u32 val, u32 mode)
 {
-	u32* ctr32 = (u32*)ctr;
+	u32 *ctr32 = (u32*)ctr;
 	
 	int i;
 	if(mode & AES_INPUT_BE)
@@ -125,9 +125,9 @@ void aes_advctr(void* ctr, u32 val, u32 mode)
 	}
 }
 
-void aes_change_ctrmode(void* ctr, u32 fromMode, u32 toMode)
+void aes_change_ctrmode(void *ctr, u32 fromMode, u32 toMode)
 {
-	u32* ctr32 = (u32*)ctr;
+	u32 *ctr32 = (u32 *)ctr;
 	int i;
 	if((fromMode ^ toMode) & AES_CNT_INPUT_ENDIAN)
 	{
@@ -147,13 +147,13 @@ void aes_change_ctrmode(void* ctr, u32 fromMode, u32 toMode)
 	}
 }
 
-void aes_batch(void* dst, const void* src, u32 blockCount)
+void aes_batch(void *dst, const void *src, u32 blockCount)
 {
 	*REG_AESBLKCNT = blockCount << 16;
 	*REG_AESCNT |=	AES_CNT_START;
 	
-	const u32* src32	= (const u32*)src;
-	u32* dst32			= (u32*)dst;
+	const u32 *src32	= (const u32 *)src;
+	u32 *dst32			= (u32 *)dst;
 	
 	u32 wbc = blockCount;
 	u32 rbc = blockCount;
@@ -180,7 +180,7 @@ void aes_batch(void* dst, const void* src, u32 blockCount)
 	}
 }
 
-void aes(void* dst, const void* src, u32 blockCount, void* iv, u32 mode, u32 ivMode)
+void aes(void *dst, const void *src, u32 blockCount, void *iv, u32 mode, u32 ivMode)
 {
 	*REG_AESCNT =	mode |
 					AES_CNT_INPUT_ORDER | AES_CNT_OUTPUT_ORDER |
@@ -233,14 +233,14 @@ u8 key2[0x10] = {
 };
 
 //Get Nand CTR key
-void getNandCTR(u8 *buf, u8 console){
-    u8 *addr = (console ? (u8*)0x080D8BBC : (u8*)0x080D797C) + 0x0F;
+void getNandCTR(u8 *buf, u32 console){
+    u8 *addr = (console ? (u8 *)0x080D8BBC : (u8 *)0x080D797C) + 0x0F;
     for(u8 keyLen = 0x10; keyLen; keyLen--)
         *(buf++) = *(addr--);
 }
 
 //Read firm0 from NAND and write to buffer
-void nandFirm0(u8 *outbuf, u32 size, u8 console){
+void nandFirm0(u8 *outbuf, u32 size, u32 console){
     u8 CTR[0x10];
     getNandCTR(CTR, console);
     aes_advctr(CTR, 0x0B130000/0x10, AES_INPUT_BE | AES_INPUT_NORMAL);
@@ -250,7 +250,7 @@ void nandFirm0(u8 *outbuf, u32 size, u8 console){
 }
 
 //Decrypts the N3DS arm9bin
-void decArm9Bin(void *armHdr, u8 mode){
+void decArm9Bin(void *armHdr, u32 mode){
 
     //Firm keys
     u8 keyY[0x10];
@@ -295,6 +295,6 @@ void setKeyXs(void *armHdr){
     for(u8 slot = 0x19; slot < 0x20; slot++){
         aes(decKey, keyData, 1, NULL, AES_ECB_DECRYPT_MODE, 0);
         aes_setkey(slot, decKey, AES_KEYX, AES_INPUT_BE | AES_INPUT_NORMAL);
-        *(u8*)(keyData+0xF) += 1;
+        *(u8 *)(keyData+0xF) += 1;
     }
 }
