+++ source/crypto.h
@@ -49,8 +49,8 @@
 #define AES_KEYY				2
 
 //NAND/FIRM stuff
-void nandFirm0(u8 *outbuf, u32 size, u8 console);
-void decArm9Bin(void *armHdr, u8 mode);
+void nandFirm0(u8 *outbuf, u32 size, u32 console);
+void decArm9Bin(void *armHdr, u32 mode);
 void setKeyXs(void *armHdr);
 
 #endif
