+++ source/fs.c
@@ -7,12 +7,12 @@
 
 static FATFS fs;
 
-u8 mountSD(void){
+u32 mountSD(void){
     if(f_mount(&fs, "0:", 1) != FR_OK) return 0;
     return 1;
 }
 
-u8 fileRead(u8 *dest, const char *path, u32 size){
+u32 fileRead(u8 *dest, const char *path, u32 size){
     FRESULT fr;
     FIL fp;
     unsigned int br = 0;
@@ -27,7 +27,7 @@ u8 fileRead(u8 *dest, const char *path, u32 size){
     return fr ? 0 : 1;
 }
 
-u8 fileWrite(const u8 *buffer, const char *path, u32 size){
+u32 fileWrite(const u8 *buffer, const char *path, u32 size){
     FRESULT fr;
     FIL fp;
     unsigned int br = 0;
@@ -39,7 +39,7 @@ u8 fileWrite(const u8 *buffer, const char *path, u32 size){
     return fr ? 0 : 1;
 }
 
-u32 fileSize(const char* path){
+u32 fileSize(const char *path){
     FIL fp;
     u32 size = 0;
 
@@ -50,7 +50,7 @@ u32 fileSize(const char* path){
     return size;
 }
 
-u8 fileExists(const char* path){
+u32 fileExists(const char *path){
     FIL fp;
     u8 exists = 0;
